--- TimberMan-master/images.qrc
@@ -1,10 +1,0 @@
-<RCC>
-    <qresource prefix="/">
-        <file>images/drovosek1_lev.png</file>
-        <file>images/drovosek1_prav.png</file>
-        <file>images/drovosek2_lev.png</file>
-        <file>images/drovosek2_prav.png</file>
-        <file>images/drovosek3_lev.png</file>
-        <file>images/drovosek3_prav.png</file>
-    </qresource>
-</RCC>
