+++ mainwindow.h
@@ -2,6 +2,8 @@
 #define MAINWINDOW_H
 
 #include <QMainWindow>
+#include <timberman.h>
+#include <tree.h>
 
 QT_BEGIN_NAMESPACE
 namespace Ui { class MainWindow; }
@@ -15,7 +17,18 @@ public:
     MainWindow(QWidget *parent = nullptr);
     ~MainWindow();
 
+    void keyPressEvent(QKeyEvent *event);
+    void cut_left();
+    void cut_right();
+    void paintEvent(QPaintEvent *event);
+    void ProvGameOver();
+
 private:
     Ui::MainWindow *ui;
+    Timberman *timberman;
+    Tree *tree=nullptr;
+    bool LeftORRight=true;
+    int gameOver=0;
+    const int interval=20;
 };
 #endif // MAINWINDOW_H
