+++ picture.qrc
@@ -1,0 +1,18 @@
+<RCC>
+    <qresource prefix="/tree">
+        <file>Images/derevo1.png</file>
+        <file>Images/derevo1_lev.png</file>
+        <file>Images/derevo1_prav.png</file>
+    </qresource>
+    <qresource prefix="/Timber">
+        <file>Images/drovosek1_lev.png</file>
+        <file>Images/drovosek1_prav.png</file>
+        <file>Images/drovosek2_lev.png</file>
+        <file>Images/drovosek2_prav.png</file>
+        <file>Images/drovosek3_lev.png</file>
+        <file>Images/drovosek3_prav (1).png</file>
+    </qresource>
+    <qresource prefix="/fon">
+        <file>Images/fon.png</file>
+    </qresource>
+</RCC>
